+++ hydromt_wflow/wflow.py
@@ -3,7 +3,7 @@
 
 import os
 from os.path import join, dirname, basename, isfile, isdir
-from typing import Union, Optional
+from typing import Union, Optional, List
 import glob
 import numpy as np
 import pandas as pd
@@ -242,7 +242,7 @@ class WflowModel(Model):
         river_upa=30,
         rivdph_method="powlaw",
         slope_len=2e3,
-        min_rivlen_ratio=0.1,
+        min_rivlen_ratio=0.0,
         min_rivdph=1,
         min_rivwth=30,
         smooth_len=5e3,
@@ -256,7 +256,9 @@ class WflowModel(Model):
         `river_upa` [km2].
 
         The river length is defined as the distance from the subgrid outlet pixel to
-        the next upstream subgrid outlet pixel.
+        the next upstream subgrid outlet pixel. The `min_rivlen_ratio` is the minimum
+        global river length to avg. cell resolution ratio and is used as a threshold in
+        window based smoothing of river length.
 
         The river slope is derived from the subgrid elevation difference between pixels at a
         half distance `slope_len` [m] up- and downstream from the subgrid outlet pixel.
@@ -301,7 +303,10 @@ class WflowModel(Model):
         slope_len : float
             length over which the river slope is calculated [km]
         min_rivlen_ratio: float
-            minimum global river length to avg. cell resolution ratio, by default 0.1
+            Ratio of cell resolution used minimum length threshold in a moving
+            window based smoothing of river length, by default 0.0
+            The river length smoothing is skipped if min_riverlen_ratio = 0.
+            For details about the river length smoothing, see :py:meth:`pyflwdir.FlwdirRaster.smooth_rivlen`
         rivdph_method : {'gvf', 'manning', 'powlaw'}
             see py:meth:`hydromt.workflows.river_depth` for details, by default "powlaw"
         smooth_len : float, optional
@@ -1380,6 +1385,84 @@ class WflowModel(Model):
             da_param = da_param.rename(key)
             self.set_staticmaps(da_param)
 
+    def setup_staticmaps_from_raster(
+        self,
+        raster_fn: str,
+        reproject_method: str,
+        variables: Optional[List] = None,
+        wflow_variables: Optional[List] = None,
+        fill_method: Optional[str] = None,
+    ) -> List[str]:
+        """
+        This component adds data variable(s) from ``raster_fn`` to staticmaps object.
+        If raster is a dataset, all variables will be added unless ``variables`` list is specified.
+        The config toml can also be updated to include the new maps using ``wflow_variables``.
+
+        Adds model layers:
+        * **raster.name** or **variables** staticmaps: data from raster_fn
+
+        Parameters
+        ----------
+        raster_fn: str
+            Source name of raster data in data_catalog.
+        reproject_method: str
+            Reprojection method from rasterio.enums.Resampling.
+            Available methods: ['nearest', 'bilinear', 'cubic', 'cubic_spline', 'lanczos', 'average', 'mode',
+            'gauss', 'max', 'min', 'med', 'q1', 'q3', 'sum', 'rms']
+        variables: list, optional
+            List of variables to add to staticmaps from raster_fn. By default all.
+        wflow_variables: list, optional
+            List of corresponding wflow variables to update the config toml (e.g: ["input.vertical.altitude"]).
+            Should match the variables list. variables list should be provided unless raster_fn contains
+            a single variable (len 1).
+        fill_method : str, optional
+            If specified, fills nodata values using fill_nodata method.
+            Available methods are {'linear', 'nearest', 'cubic', 'rio_idw'}.
+
+        Returns
+        -------
+        list
+            Names of added model staticmap layers.
+        """
+        self.logger.info(f"Preparing staticmaps data from raster source {raster_fn}")
+        # Read raster data and select variables
+        ds = self.data_catalog.get_rasterdataset(
+            raster_fn,
+            geom=self.region,
+            buffer=2,
+            variables=variables,
+            single_var_as_array=False,
+        )
+        # Fill nodata
+        if fill_method is not None:
+            ds = ds.raster.interpolate_na(method=fill_method)
+        # Reprojection
+        ds_out = ds.raster.reproject_like(self.staticmaps, method=reproject_method)
+        # Add to staticmaps
+        self.set_staticmaps(ds_out)
+
+        # Update config
+        if wflow_variables is not None:
+            self.logger.info(
+                f"Updating the config for wflow_variables: {wflow_variables}"
+            )
+            if variables is None:
+                if len(ds_out.data_vars) == 1:
+                    variables = list(ds_out.data_vars.keys())
+                else:
+                    raise ValueError(
+                        f"Cannot update the toml if raster_fn has more than one variable and variables list is not provided."
+                    )
+
+            # Check on len
+            if len(wflow_variables) != len(variables):
+                raise ValueError(
+                    f"Length of variables {variables} do not match wflow_variables {wflow_variables}. Cannot update the toml."
+                )
+            else:
+                for i in range(len(variables)):
+                    self.set_config(wflow_variables[i], variables[i])
+
     def setup_precip_forcing(
         self,
         precip_fn: str = "era5",
@@ -2101,7 +2184,6 @@ class WflowModel(Model):
         """Returns a river geometry as a geopandas.GeoDataFrame. If available, the
         stream order and upstream area values are added to the geometry properties.
         """
-        # import pdb; pdb.set_trace()
         if "rivers" in self.staticgeoms:
             gdf = self.staticgeoms["rivers"]
         elif self._MAPS["rivmsk"] in self.staticmaps: